Update of the Reusable Table Heading Component
